requirements.txt
flask
flask-cors
transformers
torch
numpy
faiss-cpu
firebase-admin
numpy